reorganizing folder structure | adding loading component
