+++ src/components/Loading/Loading.component.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './Loading.styles.css';
+
+export const Loading = (props) => {
+    return (
+        <div className="loading-page">
+            <div className ="page-loader animate-flicker">{props.message}</div>
+        </div>      
+    );
+}
