+++ src/components/SearchFilter/SearchFilter.component.jsx
@@ -1,11 +1,14 @@
 import React from 'react';
 
+import './SearchFilter.styles.css';
+
 export const SearchFilter = props => {
 	return (
-			<div>
+			<div className="custom-searchfilter">
 				<input
+					className="custom-input"
 					type="search" 
-					placeholder="busque aqui a criptomoeda"
+					placeholder="busque aqui a criptomoeda desejada..."
 					onChange={props.searchChange}
 				/>
 			</div>
