+++ src/components/SearchFilter/SearchFilter.styles.css
@@ -1,0 +1,17 @@
+.custom-searchfilter {
+    box-shadow: 0 0 4px 4px rgba(0,0,0,.3);
+    background-color: hsla(0,0%,100%,.3);
+    margin: 1rem;
+    padding: .5rem;
+    border-radius: .5rem;
+}
+
+.custom-input {
+    background-color: #f0f0f0;
+    border-color: grey;
+    font-size: 1.25rem;
+    padding: .5rem;
+    width: 100%;
+    border-style: solid;
+    border-width: .125rem;
+}
